Tweaks to CompleteYourProfile2024 related to WV-320. Added certs folder to .gitignore.
