+++ .gitignore
@@ -159,6 +159,7 @@ staticfiles
 /**/migrations/*.py
 !/**/migrations/__init__.py
 srcCordova
+src/cert/*
 src/js/compileDate.js
 src/js/config.js
 src/js/config-qa.js
