+++ src/js/components/CompleteYourProfile/Step.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import Colors from '../../../common/components/Style/Colors';
+import Colors from '../../common/components/Style/Colors';
 import StepIcon from './StepIcon';
 
 const Step = ({ onClick, step, label, completed, width }) => (
