+++ src/js/components/Navigation/HeaderBackTo.jsx
@@ -275,7 +275,7 @@ class HeaderBackTo extends Component {
         {showSignInModal && (
           <Suspense fallback={<></>}>
             <SignInModal
-              signInTitle="Sign In Or Sign Up"
+              signInTitle="Sign In or Join"
               signInSubTitle=""
               toggleOnClose={this.closeSignInModal}
               uponSuccessfulSignIn={this.closeSignInModal}
