+++ src/js/components/Navigation/HeaderBackToBallot.jsx
@@ -804,7 +804,7 @@ class HeaderBackToBallot extends Component {
           {showSignInModal && (
             <Suspense fallback={<></>}>
               <SignInModal
-                signInTitle="Sign In Or Sign Up"
+                signInTitle="Sign In or Join"
                 signInSubTitle=""
                 toggleOnClose={this.closeSignInModal}
                 uponSuccessfulSignIn={this.closeSignInModal}
