+++ src/js/components/Navigation/HeaderBackToVoterGuides.jsx
@@ -363,7 +363,7 @@ class HeaderBackToVoterGuides extends Component {
         {showSignInModal && (
           <Suspense fallback={<></>}>
             <SignInModal
-              signInTitle="Sign In Or Sign Up"
+              signInTitle="Sign In or Join"
               signInSubTitle=""
               toggleOnClose={this.closeSignInModal}
               uponSuccessfulSignIn={this.closeSignInModal}
