+++ src/js/components/Navigation/WelcomeAppbar.jsx
@@ -463,7 +463,7 @@ class WelcomeAppbar extends Component {
         {showSignInModal && (
           <Suspense fallback={<></>}>
             <SignInModal
-              signInTitle="Sign In Or Sign Up"
+              signInTitle="Sign In or Join"
               signInSubTitle=""
               toggleOnClose={this.closeSignInModal}
               uponSuccessfulSignIn={this.closeSignInModal}
